seperate wish controller
